+++ src/main/java/com/task/crudapplication/Controllers/MembershipController.java
@@ -1,42 +1,64 @@
 package com.task.crudapplication.Controllers;
 
+import com.task.crudapplication.DTOs.MembershipDto;
+import com.task.crudapplication.Entity.Book;
 import com.task.crudapplication.Entity.Membership;
+import com.task.crudapplication.Entity.User;
+import com.task.crudapplication.Exceptions.MembershipNotFound;
 import com.task.crudapplication.Services.MembershipService;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
-@RequestMapping("/api")
+@RequestMapping("/memberships")
 public class MembershipController {
     @Autowired
     MembershipService membershipService;
-    @PostMapping("/addMembership")
-    public Membership addMemberShip(@RequestBody Membership membership)
+    @PostMapping()
+    public ResponseEntity<Membership> addMemberShip(@RequestBody @Valid MembershipDto membershipDto)
     {
-        this.membershipService.addMembership(membership);
-        return membership;
+       return new ResponseEntity<>(this.membershipService.addMembership(membershipDto), HttpStatus.CREATED);
+
     }
-    @GetMapping("/getAllMembership")
-    public List<Membership> getAllMemberships(){
-       List<Membership> membershipList=(List<Membership>) this.membershipService.getAllMemberships();
-       return membershipList;
+    @GetMapping("/")
+    public ResponseEntity<List<MembershipDto>> getAllMemberships(MembershipDto membershipDto){
+       List<MembershipDto>memberships=membershipService.getAllMemberships();
+       if (memberships.size()<=0){
+           return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+       }
+       return ResponseEntity.of(Optional.of(memberships));
     }
-    @GetMapping (value = "/getMembershipById/{id}")
-    private Membership getUser(@PathVariable("id") Long id){
-        return membershipService.getMembershipById(id);
+    @GetMapping (value = "/{membershipId}")
+    private ResponseEntity <Membership> getUser(@PathVariable("membershipId") Long id) throws MembershipNotFound{
+        Membership memberships=membershipService.getMembershipById(id);
+        if (memberships==null){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }else
+            throw new MembershipNotFound("Membership not found with the id:"+id);
+
+        //return ResponseEntity.of(Optional.of(memberships));
     }
 
-    @PutMapping("/updateMembershipById/{id}")
+    @PutMapping("/{membershipId}")
     public Membership updateMembership(@RequestBody Membership membership,@PathVariable Long id){
         membership.setId(id);
         this.membershipService.updatedMembership(membership,id);
         return membership;
     }
 
-    @DeleteMapping("deleteMembershipById/{id}")
-    public String deleteMembership(@PathVariable ("id") Long id){
+    @DeleteMapping("/{membershipId}")
+    public String deleteMembership(@PathVariable ("membershipId") Long id) throws MembershipNotFound {
+        Membership membership=membershipService.getMembershipById(id);
+        if (membership==null){
+            return "Membership not found";
+        } else
          this.membershipService.deleteMembership(id);
         return "deleted successfully";
     }
